Organised utility functions by subject and removed use-cases
