+++ src/comparison/compare.ts
@@ -1,8 +1,8 @@
 import type { Change } from "diff"
 import isEqual from "lodash.isequal"
+import { xmlOutputDiff, xmlOutputMatches } from "src/comparison/xmlOutputComparison"
 import CoreHandler from "src/index"
 import convertAhoToXml from "src/lib/generateLegacyAhoXml"
-import { xmlOutputDiff, xmlOutputMatches } from "src/lib/xmlOutputComparison"
 import parseAhoXml from "src/parse/parseAhoXml"
 import type Exception from "src/types/Exception"
 import type { Trigger } from "src/types/Trigger"
