+++ src/comparison/xmlOutputComparison.ts
