+++ src/enrichAho/enrichFunctions/enrichCourtCases/offenceMatcher/offenceHasFinalResult.ts
@@ -1,5 +1,5 @@
+import resultCodeIsFinal from "src/lib/result/resultCodeIsFinal"
 import type { PncOffence } from "src/types/PncQueryResult"
-import resultCodeIsFinal from "src/use-cases/resultCodeIsFinal"
 
 const offenceHasFinalResult = (offence: PncOffence): boolean => {
   return !!offence.disposals?.some((disposal) => resultCodeIsFinal(disposal.type))
