+++ src/enrichAho/enrichFunctions/enrichCourtCases/offenceMatcher/offenceIsBreach.ts
@@ -1,5 +1,5 @@
+import getOffenceCode from "src/lib/offence/getOffenceCode"
 import type { Offence } from "src/types/AnnotatedHearingOutcome"
-import getOffenceCode from "src/utils/offence/getOffenceCode"
 
 // These are empty in Bichard, but the logic is still valid
 // (We might want to add offence codes here in the future)
