+++ src/enrichAho/enrichFunctions/enrichCourtCases/offenceMatcher/offencesAreEqual.ts
@@ -1,5 +1,5 @@
+import getOffenceCode from "src/lib/offence/getOffenceCode"
 import type { Offence } from "src/types/AnnotatedHearingOutcome"
-import getOffenceCode from "src/utils/offence/getOffenceCode"
 import offenceIsBreach from "./offenceIsBreach"
 import { offencesHaveEqualResults } from "./resultsAreEqual"
