+++ src/enrichAho/enrichFunctions/enrichOffences/index.ts
@@ -1,19 +1,18 @@
 import type { OffenceCode } from "@moj-bichard7-developers/bichard7-next-data/types/types"
 import enrichOffence from "src/enrichAho/enrichFunctions/enrichOffences/enrichOffence"
+import createCriminalProsecutionRef from "src/lib/offence/createCriminalProsecutionRef"
+import getAreaCode from "src/lib/offence/getAreaCode"
+import getOffenceCode from "src/lib/offence/getOffenceCode"
+import isOffenceCode from "src/lib/offence/isOffenceCode"
+import isOffenceIgnored from "src/lib/offence/isOffenceIgnored"
+import lookupOffenceCode from "src/lib/offence/lookupOffenceCode"
+import parseOffenceReason from "src/lib/offence/parseOffenceReason"
 import parseASN from "src/lib/parseASN"
-import parseOffenceReason from "src/lib/parseOffenceReason"
-import { ExceptionCode } from "src/types/ExceptionCode"
-import { isLookupOffenceCodeError } from "src/types/LookupOffenceCodeError"
-import isOffenceIgnored from "src/use-cases/isOffenceIgnored"
-import createCriminalProsecutionRef from "src/utils/offence/createCriminalProsecutionRef"
-import getAreaCode from "src/utils/offence/getAreaCode"
-import getOffenceCode from "src/utils/offence/getOffenceCode"
-import isOffenceCode from "src/utils/offence/isOffenceCode"
-import lookupOffenceCode from "src/utils/offence/lookupOffenceCode"
-
 import type { AnnotatedHearingOutcome, Offence } from "src/types/AnnotatedHearingOutcome"
 import type { EnrichAhoFunction } from "src/types/EnrichAhoFunction"
 import type Exception from "src/types/Exception"
+import { ExceptionCode } from "src/types/ExceptionCode"
+import { isLookupOffenceCodeError } from "src/types/LookupOffenceCodeError"
 import handle100Offences from "./handle100Offences"
 
 const enrichOffences: EnrichAhoFunction = (hearingOutCome: AnnotatedHearingOutcome) => {
