+++ src/exceptions/HO100300.ts
@@ -2,12 +2,12 @@ import { lookupOrganisationUnitByCode } from "src/dataLookup"
 import convertAsnToOrganisationUnit from "src/lib/convertAsnToOrganisationUnit"
 import errorPaths from "src/lib/errorPaths"
 import findException from "src/lib/findException"
+import populateOrganisationUnitFields from "src/lib/organisationUnit/populateOrganisationUnitFields"
 import type { OrganisationUnitCodes } from "src/types/AnnotatedHearingOutcome"
 import type Exception from "src/types/Exception"
 import type { ExceptionPath } from "src/types/Exception"
 import { ExceptionCode } from "src/types/ExceptionCode"
 import type { ExceptionGenerator } from "src/types/ExceptionGenerator"
-import populateOrganisationUnitFields from "src/use-cases/populateOrganisationUnitFields"
 
 const COURT_HEARING_LOCATION_PATH: ExceptionPath =
   "AnnotatedHearingOutcome.HearingOutcome.Hearing.CourtHearingLocation.OrganisationUnitCode".split(".")
