+++ src/exceptions/HO100305.ts
@@ -1,6 +1,6 @@
 import errorPaths from "src/lib/errorPaths"
 import { ResultClass } from "src/lib/properties"
-import resultCodeIsOnStopList from "src/lib/resultCodeIsOnStopList"
+import resultCodeIsOnStopList from "src/lib/result/resultCodeIsOnStopList"
 import type Exception from "src/types/Exception"
 import { ExceptionCode } from "src/types/ExceptionCode"
 import type { ExceptionGenerator } from "src/types/ExceptionGenerator"
