+++ src/index.ts
@@ -3,10 +3,10 @@ import generateExceptions from "./exceptions/generate"
 import parseAhoXml from "./parse/parseAhoXml"
 import parseSpiResult from "./parse/parseSpiResult"
 import transformSpiToAho from "./parse/transformSpiToAho"
+import generateTriggers from "./triggers/generate"
 import type { AnnotatedHearingOutcome } from "./types/AnnotatedHearingOutcome"
 import type BichardResultType from "./types/BichardResultType"
 import type PncGateway from "./types/PncGateway"
-import generateTriggers from "./use-cases/generateTriggers"
 
 export default (message: string, pncGateway: PncGateway): BichardResultType => {
   let hearingOutcome: AnnotatedHearingOutcome | Error
