+++ src/lib/offence/getOffenceFullCode.ts
