+++ src/lib/offence/parseOffenceReason.ts
@@ -1,6 +1,6 @@
+import { getLocalOffenceReason, getNationalOffenceReason } from "src/lib/offence/getOffenceReason"
+import regexTestNationalOffenceCode from "src/lib/offence/regexTestNationalOffenceCode"
 import type { OffenceReason } from "src/types/AnnotatedHearingOutcome"
-import { getLocalOffenceReason, getNationalOffenceReason } from "src/utils/offence/getOffenceReason"
-import regexTestNationalOffenceCode from "src/utils/offence/regexTestNationalOffenceCode"
 
 // parse an offence code into an national offence reason or a local offence reason if the national reason isn't valid
 export default (offenceCode: string, areaCode: string, reason?: OffenceReason): OffenceReason | undefined => {
