+++ src/lib/result/isAdjourned.ts
@@ -1,4 +1,4 @@
-import { ADJOURNMENT_RANGES } from "./properties"
+import { ADJOURNMENT_RANGES } from "src/lib/properties"
 
 const isAdjourned = (cjsResultCode?: number) =>
   cjsResultCode && ADJOURNMENT_RANGES.some((range) => cjsResultCode >= range[0] && cjsResultCode <= range[1])
