+++ src/lib/result/isWarrantIssued.ts
@@ -1,4 +1,4 @@
-import { WARRANT_ISSUED_CODES } from "./properties"
+import { WARRANT_ISSUED_CODES } from "src/lib/properties"
 
 const isWarrantIssued = (cjsResultCode?: number) =>
   cjsResultCode && WARRANT_ISSUED_CODES.some((range) => cjsResultCode >= range[0] && cjsResultCode <= range[1])
