+++ src/lib/result/resultCodeIsFinal.ts
@@ -1,0 +1,6 @@
+import { lookupResultCodeByCjsCode } from "src/dataLookup"
+
+const resultCodeIsFinal = (resultCode: number): boolean =>
+  lookupResultCodeByCjsCode(resultCode.toString())?.type === "F"
+
+export default resultCodeIsFinal
