+++ src/lib/result/resultCodeIsOnStopList.ts
@@ -1,4 +1,4 @@
-import { STOP_LIST } from "./properties"
+import { STOP_LIST } from "src/lib/properties"
 
 const resultCodeIsOnStopList = (code: number): boolean => STOP_LIST.includes(code)
