+++ src/parse/transformSpiToAho/PopulateOffences.ts
@@ -11,7 +11,7 @@ import {
   INDICTMENT,
   TIME_RANGE
 } from "src/lib/properties"
-import resultCodeIsOnStopList from "src/lib/resultCodeIsOnStopList"
+import resultCodeIsOnStopList from "src/lib/result/resultCodeIsOnStopList"
 import type { CriminalProsecutionReference, Offence, OffenceCode } from "src/types/AnnotatedHearingOutcome"
 import type { OffenceParsedXml, ResultedCaseMessageParsedXml, SpiResult } from "src/types/SpiResult"
 import removeSeconds from "src/utils/removeSeconds"
