+++ src/parse/transformSpiToAho/populateHearing.ts
@@ -1,6 +1,6 @@
+import getOrganisationUnit from "src/lib/organisationUnit/getOrganisationUnit"
 import type { Hearing } from "src/types/AnnotatedHearingOutcome"
 import type { ResultedCaseMessageParsedXml } from "src/types/SpiResult"
-import getOrganisationUnit from "src/use-cases/getOrganisationUnit"
 import removeSeconds from "src/utils/removeSeconds"
 
 export default (messageId: string, courtResult: ResultedCaseMessageParsedXml): Hearing => {
