+++ src/triggers/TRPR0004.ts
@@ -1,9 +1,9 @@
+import getOffenceFullCode from "src/lib/offence/getOffenceFullCode"
 import type { Offence } from "src/types/AnnotatedHearingOutcome"
 import type { Trigger } from "src/types/Trigger"
 import { TriggerCode } from "src/types/TriggerCode"
 import type { TriggerGenerator } from "src/types/TriggerGenerator"
 import { CjsVerdict } from "src/types/Verdict"
-import getOffenceFullCode from "src/utils/offence/getOffenceFullCode"
 
 const triggerCode = TriggerCode.TRPR0004
 const resultCodes = [3052, 3081, 3085, 3086, 3087, 3088, 3089, 3090, 3091, 1179, 1181, 3281, 3282]
