+++ src/triggers/TRPR0025.ts
@@ -1,7 +1,7 @@
+import getOffenceFullCode from "src/lib/offence/getOffenceFullCode"
 import type { AnnotatedHearingOutcome, Offence } from "src/types/AnnotatedHearingOutcome"
 import { TriggerCode } from "src/types/TriggerCode"
 import type { TriggerGenerator } from "src/types/TriggerGenerator"
-import getOffenceFullCode from "src/utils/offence/getOffenceFullCode"
 
 const triggerCode = TriggerCode.TRPR0025
 const validMatches = [
