+++ src/triggers/TRPR0030.ts
@@ -1,8 +1,8 @@
 import isCaseRecordable from "src/lib/isCaseRecordable"
+import getOffenceFullCode from "src/lib/offence/getOffenceFullCode"
 import type { Offence } from "src/types/AnnotatedHearingOutcome"
 import { TriggerCode } from "src/types/TriggerCode"
 import type { TriggerGenerator } from "src/types/TriggerGenerator"
-import getOffenceFullCode from "src/utils/offence/getOffenceFullCode"
 
 const triggerCode = TriggerCode.TRPR0030
 const offenceCodes = ["PL84504", "PL84505", "PL84506"]
