--- src/use-cases/resultCodeIsFinal.ts
@@ -1,5 +1,0 @@
-import findResultCode from "./findResultCode"
-
-const resultCodeIsFinal = (resultCode: number): boolean => findResultCode(resultCode).type === "F"
-
-export default resultCodeIsFinal
